Adding the first program; Cramers Rule practice
While there are other spots to start programming for this course, this was the first spot where it became imperative due to the wide general use of this one particular method in its easier stages of 2x2 and 3x3 matrices that a Circuits course can infinitely operate around.

Prior to this point, there are several general forms of a multitude of basic circuit designs. As such, while I could focus on specific examples and grow cases over time, this is something to work on later.